Reorganize Phase 1 files into extractor/ subfolder; fix relative imports
