+++ a_quantale_theoretic_approach/extractor/gnn_truth_value.py
