+++ a_quantale_theoretic_approach/extractor/gnn_weights.pth
